+++ app/src/main/java/sugar/free/sightremote/database/PumpStatusChanged.java
@@ -16,10 +16,10 @@ import sugar.free.sightparser.applayer.descriptors.PumpStatus;
 public class PumpStatusChanged {
 
     @DatabaseField(generatedId = true)
-    private int id;
+    private long id;
 
     @DatabaseField
-    private int eventNumber;
+    private long eventNumber;
 
     @DatabaseField
     private String pump;
